+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ## Unreleased
 
+- Added HDF5 support for `nsReLU` layer.
+
 ## 5.2.0
 
 - Added AIS support for `nsReLU` layer.